[FIX] Fix issues and duplicate code
